requirements.txt
alembic==1.12.1
async-timeout==4.0.3
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
click==8.1.7
Deprecated==1.2.14
Flask==2.3.3
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-JWT==0.3.2
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
psycopg2-binary==2.9.9
Pygments==2.16.1
PyJWT==1.4.2
python-dotenv==1.0.0
redis==5.0.1
rich==13.6.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
